refactor?: Refact and organize project
