+++ core.js
@@ -1,0 +1,6 @@
+export { SemiReactive } from "./src/core/main.js";
+export { Component } from "./src/core/component.js";
+export { Switch } from "./src/core/switch.js";
+export { Injectable } from "./src/core/injectable.js";
+export { Service } from "./src/core/service.js";
+export { Request } from "./src/core/request.js";
