+++ src/core/property.js
@@ -1,7 +1,6 @@
-import { Subject } from "./rx.js";
+import { Subject } from "../../rx.js";
 
-export class Property
-{
+export class Property {
 	get value() {
 		return this.__value;
 	}
@@ -26,7 +25,7 @@ export class Property
 		if (component !== null && component !== undefined) {
 			this.change = () => component.reload();
 		} else {
-			this.change = () => {};
+			this.change = () => { };
 		}
 	}
 
@@ -40,32 +39,4 @@ export class Property
 			};
 		}
 	}
-}
-
-export class PropertySet
-{
-    constructor() {
-        this.properties = {};
-    }
-
-    set(key, property) {
-        this.properties[key] = property;
-    }
-
-    get(key) {
-        return this.properties[key];
-    }
-
-    keys() {
-        return Object.keys(this.properties);
-    }
-
-    getAll() {
-        const ret = [];
-
-        for (let key of this.keys()) {
-            ret.push({ key, property: this.get(key) });
-        }
-        return ret;
-    }
 }
