+++ src/core/service.js
@@ -1,0 +1,19 @@
+import { Subject } from "../../rx.js";
+
+export class Service {
+  constructor() {
+    this.events$ = new Subject();
+  }
+
+  onRegister() { }
+
+  onGet() { }
+
+  notify(event) {
+    this.events$.next(event);
+  }
+
+  events() {
+    return this.events$;
+  }
+}
