+++ src/navigation/router.js
@@ -1,4 +1,4 @@
-import { Switch } from "../core/components.js";
+import { Switch } from "../../core.js";
 
 export class Router extends Switch
 {
@@ -65,40 +65,3 @@ export class Router extends Switch
 		}
 	}
 }
-
-export class HashRouter extends Router
-{
-	constructor(props) {
-		super(props);
-		this.__init = false;
-	}
-
-	init() {
-		window.addEventListener("hashchange", () => this.loadPath());
-		this.__init = true;
-	}
-
-	setRoutes(routes) {
-		super.setRoutes(routes);
-
-		if (!this.__init) {
-			this.init();
-		}
-	}
-
-	getUrlPath() {
-		const s_ret = super.getUrlPath();
-
-		if (s_ret === false) {
-			const hash = window.location.hash.replace(/^#\//, '');
-			return hash.split('/');
-		}
-
-		return s_ret;
-	}
-}
-
-/**
- * @deprecated - Use HashRouter
- */
-export const SimpleRouter = HashRouter;
