+++ src/rx/subscription.js
@@ -1,0 +1,13 @@
+export class Subscription {
+	constructor() {
+		this.__subscriptions = [];
+	}
+
+	add(subscription) {
+		this.__subscriptions.push(subscription);
+	}
+
+	unsubscribe() {
+		this.__subscriptions.forEach(subscription => subscription.unsubscribe());
+	}
+}
