+++ src/tools/text.js
@@ -1,4 +1,4 @@
-import { Component } from "../core/components.js";
+import { Component } from "../../core.js";
 
 export class TextComponent extends Component
 {
